backend/requirements.txt
aiofiles==0.4.0
astroid==2.3.3
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
cryptography==2.8
cycler==0.10.0
Django==2.2.7
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.44.1
Flask==1.1.1
h11==0.8.1
httptools==0.0.13
idna==2.8
image==1.5.27
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.4
passlib==1.7.2
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
pydantic==1.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.5
PyQt5==5.13.2
PyQt5-sip==12.7.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-multipart==0.0.5
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
sqlparse==0.3.0
starlette==0.12.9
uvicorn==0.10.8
uvloop==0.14.0
websockets==8.1
Werkzeug==0.16.0
wrapt==1.11.2
